ipuin baten eszenak segidan, bideo moduan, ikusteko aukera. Parte 2

+++ www/js/directives/objektua.directive.js
@@ -5,20 +5,22 @@ app.directive ('objektua', ['$cordovaDialogs', 'Database', function ($cordovaDia
     scope: {},
     template : '<img class="laukia" hm-rotatestart="onRotateStart" hm-rotate="onRotate" hm-rotateend="onRotateEnd" hm-pinch="onPinch" hm-pinchend="onPinchEnd" hm-panstart="onPanStart" hm-panmove="onPan" hm-panend="onPanEnd" hm-press="onPress" ng-dblclick="onDblClick()">',
     link: function (scope, element, attrs){
+      
       var initScale = attrs.scale !== undefined ? attrs.scale : 1,
           initAngle = attrs.rotate !== undefined ? attrs.rotate : 0,
           rotationInit = 0,
           transform = {  translate :{ x: attrs.x, y: attrs.y   }, scale: initScale, angle: initAngle, rx: 0, ry: 0, rz: 0 },
           abiapuntua = {'x': 0, 'y': 0},
           limits = {'top': 0, 'right': 0, 'bottom': 0, 'left': 0},
-          eszenatokia = angular.element ('#eszenatokia');
+          eszenatokia = angular.element ('#eszenatokia'),
+          loki = attrs.lock == 'true'; // no se recibe como boolean....
           
       limits.left = eszenatokia[0].offsetLeft + 15;
       limits.top = eszenatokia[0].offsetTop + 15;
       limits.right = eszenatokia[0].offsetWidth - 15;
       limits.bottom = eszenatokia[0].offsetHeight - 15;
       
-      element.children ().attr ('src', attrs.background);
+      element.children ().attr ('src', attrs.src);
       
       // Le damos "id" al elemento para poder hacer una txapuzilla luego
       element.children ().attr ('id', 'objektua_' + element.attr ('data-eo-id'));
@@ -29,33 +31,38 @@ app.directive ('objektua', ['$cordovaDialogs', 'Database', function ($cordovaDia
       var updateElementTransform = function (transform_new, dbGorde){
         dbGorde = typeof dbGorde !== 'undefined' ? dbGorde : false;
         
-        // Aplicamos los cambios en el CSS
-        element.children ().css (transform2css (transform_new));
-        
-        // Pase lo que pase, vayas donde vayas, hagas lo que hagas, ponte bragas
-        // Comprobamos que tras los cambios el objeto no se salga de los limites
-        var bounds = document.getElementById ("objektua_" + element.attr ('data-eo-id')).getBoundingClientRect ();
-        if (bounds.top < limits.top || bounds.bottom > limits.bottom || bounds.right > limits.right || bounds.left < limits.left){
-          // Sale de los limites -> restablecemos los valores (nos quedamos como estabamos)
-          element.children ().css (transform2css (transform));
+        if (!loki){
           
-          return (false);
-        }
-        else{
-          // No se sale de los limites -> guardamos los valores nuevos y palante
-          transform = transform_new;
-        
-          if (dbGorde){
-            var id = parseInt (element.attr ('data-eo-id'));
-            var style = JSON.stringify (element[0].children[0].style);
-        
-            Database.query ('UPDATE eszena_objektuak SET style=? WHERE id=?', [style, id]).then (function (){}, function (error){
-              console.log ("Objektua directive UPDATE eszena_objektuak", error);
-            });
+          // Aplicamos los cambios en el CSS
+          element.children ().css (transform2css (transform_new));
+          
+          // Pase lo que pase, vayas donde vayas, hagas lo que hagas, ponte bragas
+          // Comprobamos que tras los cambios el objeto no se salga de los limites
+          var bounds = document.getElementById ("objektua_" + element.attr ('data-eo-id')).getBoundingClientRect ();
+          if (bounds.top < limits.top || bounds.bottom > limits.bottom || bounds.right > limits.right || bounds.left < limits.left){
+            // Sale de los limites -> restablecemos los valores (nos quedamos como estabamos)
+            element.children ().css (transform2css (transform));
+            
+            return (false);
+          }
+          else{
+            // No se sale de los limites -> guardamos los valores nuevos y palante
+            transform = transform_new;
+          
+            if (dbGorde){
+              var id = parseInt (element.attr ('data-eo-id'));
+              var style = JSON.stringify (element[0].children[0].style);
+          
+              Database.query ('UPDATE eszena_objektuak SET style=? WHERE id=?', [style, id]).then (function (){}, function (error){
+                console.log ("Objektua directive UPDATE eszena_objektuak", error);
+              });
+            }
+            
+            return (true);
           }
           
-          return (true);
         }
+        
       };
       
       var transform2css = function (t){
@@ -68,27 +75,31 @@ app.directive ('objektua', ['$cordovaDialogs', 'Database', function ($cordovaDia
       
       scope.onPress = function (){
         
-        $cordovaDialogs.confirm ('Ezabatu nahi duzu?', 'EZABATU', ['BAI', 'EZ']).then (function (buttonIndex){
+        if (!loki){
           
-          if (buttonIndex == 1){
-            
-            if (element.attr ('data-eo-id') !== undefined){
+          $cordovaDialogs.confirm ('Ezabatu nahi duzu?', 'EZABATU', ['BAI', 'EZ']).then (function (buttonIndex){
+          
+            if (buttonIndex == 1){
               
-              Database.query ('DELETE FROM eszena_objektuak WHERE id=?', [parseInt (element.attr ('data-eo-id'))]).then (function (){
+              if (element.attr ('data-eo-id') !== undefined){
                 
-                element.remove();
+                Database.query ('DELETE FROM eszena_objektuak WHERE id=?', [parseInt (element.attr ('data-eo-id'))]).then (function (){
+                  
+                  element.remove();
+                  
+                }, function (error){
+                  console.log ("Objektua directive DELETE", error);
+                });
                 
-              }, function (error){
-                console.log ("Objektua directive DELETE", error);
-              });
+              }
               
             }
             
-          }
+          }, function (error){
+            console.log ("Objektua directive onPress", error);
+          });
           
-        }, function (error){
-          console.log ("Objektua directive onPress", error);
-        });
+        }
         
       };
